backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-multipart
python-dotenv
requests
feedparser
passlib[bcrypt]
python-jose[cryptography]
argon2-cffi
resend